requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.0
azure-ai-inference
azure-core 